# Open MP in THBlas for GEMM speedup