Speed up the GEMM calculation by introducing Open MP in THBlas.
The changes have been tested on Devserver with multi-thread calculations.
To initiate the Open MP computing, buck build caffe2:ifbpy @mode/opt-lto.
Use torch.set_num_threads() to change the number of threads. If not specified, the code will use all the available cores in the server.

It's found that multi-thread computing is actually slower for cases "!transa". This is because "#pragma omp parallel" is applied to the innermost "for" loop (see branch 1 and 3). Some compilers introduce errors if I put "omp parallel" to the outermost loop in branch 1 and 3. I guess it takes more cashe memory to set up "omp parallel". I suggest to remove "omp parallel" in those two branches for now. We could investigate this later with a new task.

The number of threads used here is  1
---------------------------------------------------------------
Case !transa && !transb
loops:  1        x:     8192     y:     8192     z:     1024     time:  212278045ns
---------------------------------------------------------------
Case !transa && transb
loops:  1        x:     8192     y:     8192     z:     1024     time:  197709088ns
---------------------------------------------------------------
Case transa && !transb
loops:  1        x:     8192     y:     8192     z:     1024     time:  101073193ns
---------------------------------------------------------------
Case transa && transb
loops:  1        x:     8192     y:     8192     z:     1024     time:  1597203337ns



The number of threads used here is  8
---------------------------------------------------------------
Case !transa && !transb
loops:  1        x:     8192     y:     8192     z:     1024     time:  274187884ns     =>  acceleration multiplier: 0.77
---------------------------------------------------------------
Case !transa && transb
loops:  1        x:     8192     y:     8192     z:     1024     time:  277355210ns     =>  acceleration multiplier: 0.71
---------------------------------------------------------------
Case transa && !transb
loops:  8        x:     8192     y:     8192     z:     1024     time:  13643668ns     =>  acceleration multiplier: 7.41
---------------------------------------------------------------
Case transa && transb
loops:  1        x:     8192     y:     8192     z:     1024     time:  206398599ns     =>  acceleration multiplier: 7.74



The number of threads used here is  16
Case !transa && !transb
---------------------------------------------------------------
loops:  1        x:     8192     y:     8192     z:     1024     time:  435479635ns     =>  acceleration multiplier: 0.49
---------------------------------------------------------------
Case !transa && transb
loops:  1        x:     8192     y:     8192     z:     1024     time:  409712394ns     =>  acceleration multiplier: 0.48
---------------------------------------------------------------
Case transa && !transb
loops:  13       x:     8192     y:     8192     z:     1024     time:  8076201ns     =>  acceleration multiplier: 12.51
---------------------------------------------------------------
Case transa && transb
loops:  1        x:     8192     y:     8192     z:     1024     time:  116395570ns     =>  acceleration multiplier: 13.72