@ezyang 2019-03-14
Test failure looks legit:

```
Mar 14 22:11:25 + /var/lib/jenkins/.local/lib/python2.7/site-packages/torch/test/broadcast_test --gtest_output=xml:/var/lib/jenkins/workspace/caffe2_tests/cpp/broadcast_test.xml
Mar 14 22:11:25 Running main() from /var/lib/jenkins/workspace/third_party/googletest/googletest/src/gtest_main.cc
Mar 14 22:11:25 [==========] Running 1 test from 1 test case.
Mar 14 22:11:25 [----------] Global test environment set-up.
Mar 14 22:11:25 [----------] 1 test from BroadcastTest
Mar 14 22:11:25 [ RUN      ] BroadcastTest.Broadcast
Mar 14 22:11:28 /var/lib/jenkins/workspace/aten/src/ATen/test/broadcast_test.cpp:143: Failure
Mar 14 22:11:28 Value of: a.addmm(b, c).equal(a.expand({5, 7}).addmm(b, c))
Mar 14 22:11:28   Actual: false
Mar 14 22:11:28 Expected: true
Mar 14 22:11:28 /var/lib/jenkins/workspace/aten/src/ATen/test/broadcast_test.cpp:153: Failure
Mar 14 22:11:28 Value of: aScalar.addmm(b, c).equal(aScalar.expand({5, 7}).addmm(b, c))
Mar 14 22:11:28   Actual: false
Mar 14 22:11:28 Expected: true
```