@zhangguanheng66 2019-03-19
@cpuhrsch The PR is ready. Please note:
It's found that multi-thread computing is actually slower for cases "!transa". This is because "#pragma omp parallel" is applied to the innermost "for" loop (see branch 1 and 3). Some compilers introduce errors if I put "omp parallel" to the outermost loop in branch 1 and 3. I guess it takes more cashe memory to set up "omp parallel". I suggest to remove "omp parallel" in those two branches for now. We could investigate this later with a new task.