@cpuhrsch 2019-03-19
Thanks @zhangguanheng66. Could you provide the test script you used to generate those numbers?