@zhangguanheng66 2019-03-19
@cpuhrsch  Change "num_threads" in the file to control the number of thread.

[test_mm.txt](https://github.com/pytorch/pytorch/files/2984043/test_mm.txt)