@cpuhrsch 2019-03-19
@zhangguanheng66 - Thanks :) - Could I please ask you to parameterize the values you varied for above script using argparse and post a set of command invocations to generate the benchmarks?